thoa's first attempt
My first attempt with creating a Cube class, a Pair class, and a
PairList class. Not all required methods have been implemented yet. I
wonder if Cube inherits Pair?
Most of the "heavy-duty" methods are in the PairList class. We have to
see if they are correct in terms of the math behind them and
syntax-wise. Then we'll have to see if there are better way to re-code
them.
I'm still relatively short-skilled and inexperienced in programming and
in algorithms so please bear with my code. Look forward to learning!